+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Change Log
 
-## 0.0.1
+### 0.1.0
+- Rebuilded all commands to use easy
 
+### 0.0.4
+- Added commmand "Selected Multiline Text to Selections"
+- Added command "Copy Selected Text N Times"
+
+### 0.0.3
+- Improved command name.
+- Changed behavior of join parameters.
+- Added join at a line command.
+- Added comma to line ends command.
+
+### 0.0.2
+- Added MultiLine commands.
+- Added fixed seperator commands.
+
+### 0.0.1
 - Initial release